+++ docs/html/search/functions_0.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['acetime_5fstrcmp_5fpp',['acetime_strcmp_PP',['../compat_8h.html#a286dd84f632d303baf33b0c82896b6b2',1,'compat.cpp']]],
   ['addactivecandidatestoactivepool',['addActiveCandidatesToActivePool',['../classace__time_1_1extended_1_1TransitionStorage.html#a114cb0ef4591f824fb41a81b329cded6',1,'ace_time::extended::TransitionStorage']]],
   ['addfreeagenttoactivepool',['addFreeAgentToActivePool',['../classace__time_1_1extended_1_1TransitionStorage.html#a92f9a9c0bc6880f887c9f51ecce42f4e',1,'ace_time::extended::TransitionStorage']]],
   ['addfreeagenttocandidatepool',['addFreeAgentToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorage.html#a536a8449ba143034cb74dd181f2d9340',1,'ace_time::extended::TransitionStorage']]],