+++ examples/CommandLineClock/config.h
@@ -21,7 +21,7 @@
 // ExtendedZoneProcessor at the same time.
 
 #ifndef AUNITER
-  #warning Using Arduino IDE configuration
+  // Arduino IDE in interactive mode
   #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
   #define ENABLE_TIME_ZONE_TYPE_BASIC 1
   #define ENABLE_TIME_ZONE_TYPE_EXTENDED 1
@@ -33,6 +33,10 @@
   #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
   #define ENABLE_TIME_ZONE_TYPE_BASIC 1
   #define ENABLE_TIME_ZONE_TYPE_EXTENDED 0
+#elif defined(AUNITER_SAMD)
+  #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
+  #define ENABLE_TIME_ZONE_TYPE_BASIC 1
+  #define ENABLE_TIME_ZONE_TYPE_EXTENDED 1
 #elif defined(AUNITER_ESP8266)
   #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_NTP
   #define ENABLE_TIME_ZONE_TYPE_BASIC 1