+++ examples/OledClock/config.h
@@ -9,7 +9,7 @@
 
 #define EEPROM_SIZE 32
 
-// Set to 1 to print debugging info to Serial
+// Set to 1 to print debugging info to SERIAL_PORT_MONITOR
 #define ENABLE_SERIAL 0
 
 // Set to 1 to force the ClockInfo to its initial state
@@ -26,12 +26,14 @@
 #define TIME_SOURCE_TYPE_DS3231 1
 #define TIME_SOURCE_TYPE_NTP 2
 #define TIME_SOURCE_TYPE_BOTH 3 
-#ifndef AUNITER
-  #define AUNITER_MICRO_MINDER
-  #warning Using default AUNITER_MICRO_MINDER
-#endif
 
-#if defined(AUNITER_NANO)
+#ifndef AUNITER
+  // Arduino IDE in interactive mode
+  #define MODE_BUTTON_PIN 2
+  #define CHANGE_BUTTON_PIN 3
+  #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
+  #define OLED_REMAP false
+#elif defined(AUNITER_NANO)
   #define MODE_BUTTON_PIN 2
   #define CHANGE_BUTTON_PIN 3
   #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
@@ -46,6 +48,11 @@
   #define CHANGE_BUTTON_PIN 3
   #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
   #define OLED_REMAP false
+#elif defined(AUNITER_SAMD)
+  #define MODE_BUTTON_PIN 11
+  #define CHANGE_BUTTON_PIN 10
+  #define TIME_SOURCE_TYPE TIME_SOURCE_TYPE_DS3231
+  #define OLED_REMAP true
 #elif defined(AUNITER_ESP8266) || defined(AUNITER_ESP_MINDER)
   #define MODE_BUTTON_PIN D4
   #define CHANGE_BUTTON_PIN D3