+++ src/ace_time/BasicZoneProcessor.h
@@ -6,13 +6,12 @@
 #ifndef ACE_TIME_BASIC_ZONE_PROCESSOR_H
 #define ACE_TIME_BASIC_ZONE_PROCESSOR_H
 
-#include <Arduino.h>
 #include <string.h> // strchr()
 #include <stdint.h>
 #include "internal/ZonePolicy.h"
 #include "internal/ZoneInfo.h"
 #include "internal/Brokers.h"
-#include "common/logger.h"
+#include "common/logging.h"
 #include "TimeOffset.h"
 #include "LocalDate.h"
 #include "OffsetDateTime.h"