+++ src/ace_time/ExtendedZoneProcessor.cpp
@@ -3,6 +3,7 @@
  * Copyright (c) 2019 Brian T. Park
  */
 
+#include <Print.h>
 #include "LocalDate.h"
 #include "ExtendedZone.h"
 #include "ExtendedZoneProcessor.h"