+++ src/ace_time/ExtendedZoneProcessor.h
@@ -6,12 +6,12 @@
 #ifndef ACE_TIME_EXTENDED_ZONE_PROCESSOR_H
 #define ACE_TIME_EXTENDED_ZONE_PROCESSOR_H
 
-#include <Arduino.h>
 #include <string.h> // memcpy()
 #include <stdint.h>
+#include "common/compat.h"
 #include "internal/ZonePolicy.h"
 #include "internal/ZoneInfo.h"
-#include "common/logger.h"
+#include "common/logging.h"
 #include "TimeOffset.h"
 #include "LocalDate.h"
 #include "OffsetDateTime.h"